ml-model/requirements.txt
flask
pandas
scikit-learn
joblib
python-dotenv
numpy
